requirements.txt
scrapy>=2.11.0
requests>=2.31.0
PyPDF2>=3.0.1
pdfplumber>=0.10.0
python-docx>=0.8.11
openpyxl>=3.1.2
beautifulsoup4>=4.12.2
lxml>=4.9.3
selenium>=4.15.0
fake-useragent>=1.4.0
urllib3>=2.0.0
chardet>=5.2.0
python-magic>=0.4.27
python-pptx>=0.6.21